requirements.txt

certifi==2020.6.20
chardet==3.0.4
db==0.1.1
db-sqlite3==0.0.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-phonenumber-field==5.0.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
entrypoints==0.3
filelock==3.0.12
flake8==3.8.3
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==20.0.4
idna==2.10
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.4.2
jedi==0.17.0
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
lazy-object-proxy==1.4.3
markdown2==2.3.9
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.0
pyflakes==2.2.0
pyglet==1.5.7
pyls==0.1.6
pymunk==5.6.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.32.0
pytiled-parser==0.9.4a3
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.24.0
rope==0.17.0
Send2Trash==1.5.0
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0
sqlparse==0.3.1
testpath==0.4.4
toml==0.10.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
yapf==0.30.0
